rename spark native reader
