+++ lakesoul-spark/src/main/scala/org/apache/spark/sql/execution/datasources/v2/merge/parquet/Native/NativeMergePartitionReaderFactory.scala
@@ -45,15 +45,15 @@ import scala.collection.mutable.ArrayBuffer
  * @param partitionSchema Schema of partitions.
  *                        //  * @param filterMap Filters to be pushed down in the batch scan.
  */
-case class NativeMergeParquetPartitionReaderFactory(sqlConf: SQLConf,
-                                                    broadcastedConf: Broadcast[SerializableConfiguration],
-                                                    dataSchema: StructType,
-                                                    readDataSchema: StructType,
-                                                    partitionSchema: StructType,
-                                                    filters: Array[Filter],
-                                                    mergeOperatorInfo: Map[String, MergeOperator[Any]],
-                                                    defaultMergeOp: MergeOperator[Any],
-                                                    options: Map[String, String] = Map.empty)
+case class NativeMergePartitionReaderFactory(sqlConf: SQLConf,
+                                             broadcastedConf: Broadcast[SerializableConfiguration],
+                                             dataSchema: StructType,
+                                             readDataSchema: StructType,
+                                             partitionSchema: StructType,
+                                             filters: Array[Filter],
+                                             mergeOperatorInfo: Map[String, MergeOperator[Any]],
+                                             defaultMergeOp: MergeOperator[Any],
+                                             options: Map[String, String] = Map.empty)
   extends NativeMergeFilePartitionReaderFactory(mergeOperatorInfo, defaultMergeOp) with Logging {
 
   private val isCaseSensitive = sqlConf.caseSensitiveAnalysis
