+++ lakesoul-spark/src/main/scala/org/apache/spark/sql/execution/datasources/v2/parquet/NativePartitionReaderFactory.scala
@@ -41,12 +41,12 @@ import scala.collection.mutable
   * @param partitionSchema Schema of partitions.
   *                        //  * @param filterMap Filters to be pushed down in the batch scan.
   */
-case class NativeParquetPartitionReaderFactory(sqlConf: SQLConf,
-                                               broadcastedConf: Broadcast[SerializableConfiguration],
-                                               dataSchema: StructType,
-                                               readDataSchema: StructType,
-                                               partitionSchema: StructType,
-                                               filters: Array[Filter])
+case class NativePartitionReaderFactory(sqlConf: SQLConf,
+                                        broadcastedConf: Broadcast[SerializableConfiguration],
+                                        dataSchema: StructType,
+                                        readDataSchema: StructType,
+                                        partitionSchema: StructType,
+                                        filters: Array[Filter])
   extends NativeFilePartitionReaderFactory with Logging {
   private val isCaseSensitive = sqlConf.caseSensitiveAnalysis
   private val enableOffHeapColumnVector = sqlConf.offHeapColumnVectorEnabled
